more bugs left to fix

+++ player.h
@@ -10,21 +10,21 @@ class Player: virtual protected Entity{
     public:
     
     class Enemy *E_Target;
-    void set_stats(std::string NAME, double HEALTH, double ATTACK, double DEFENSE);
+    void set_stats(std::string NAME, int HEALTH, int ATTACK, int DEFENSE);
     void take_dmg(Enemy *E_Target);
     void attack(Enemy *E_Target);
     std::string Name;
     void set_Name(std::string Name);
-    void set_Def(double Defense);
-    void set_HP(double Health);
-    void set_Atk(double Attack);
-    double Attack;
+    void set_Def(int Defense);
+    void set_HP(int Health);
+    void set_Atk(int Attack);
+    int Attack;
     private:
 
      std::string name;
-     double Def;
-     double HP;
-     double Atk;
+     int Def;
+     int HP;
+     int Atk;
 
 };
